requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
googlemaps==4.10.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
